move article summary to article
