+++ src/components/article/article-summary/ArticleSummaryCard.scss
